Fix overlapping.
